fix(main): allow a custom theme to be specified in options.theme (#31)
## Description

Before some recent updates, I was able to use

```lua
require('drop').setup({
    theme = {
        symbols = {
            -- my symbols here
        },
        colors = {
            -- my colors here
        },
    },
})
```

Now with the preset themes, this no longer works. Also, it is specified
that `options.theme` can be a `DropTheme` in addition to a string.
However, if it's not a string then it breaks.

This PR allows non-string values to be set for `options.theme`.